Separate into alpha/v1 MEDFORD files
Created new folders to separate MEDFORD files in the 'alpha' version of
MEDFORD and files that have been validated using MEDFORD version 1.0.
Also updated an example to work with MEDFORD 1.0, although since we
aren't the ones who did the research, we are missing critical
information to allow it to completely validate.