Remove large files, revert to Data-Ref from Copy
Decided that because these are just example files, we are not going to
provide copies of all of the files, as otherwise the repository would
get very large very fast. However, this is still something that should
be looked at, since we are currently hard-linking to online journal
resources, which is poor form.